create iCEBreaker VGA example
Redid calculations for timings and number of clock cycles
based on the 12MHz clock of the iCEBreaker board

This has been tested to provide stable VGA output on a few
different small VGA screens.

Output is a three solid color output.

Also created a testbench for verifying waveforms and timings.